fix(components): move json info component
