+++ src/lib/pages/contract-details/index.tsx
@@ -22,6 +22,7 @@ import { AssetsSection } from "lib/components/asset";
 import { CustomTab } from "lib/components/CustomTab";
 import { DelegationsSection } from "lib/components/delegations";
 import { CustomIcon } from "lib/components/icon";
+import { JsonInfo } from "lib/components/json/JsonInfo";
 import { Loading } from "lib/components/Loading";
 import PageContainer from "lib/components/PageContainer";
 import { ErrorFetching, InvalidState } from "lib/components/state";
@@ -34,7 +35,6 @@ import { ContractDesc } from "./components/contract-description";
 import { ContractStates } from "./components/contract-states";
 import { ContractTop } from "./components/ContractTop";
 import { InstantiateInfo } from "./components/InstantiateInfo";
-import { JsonInfo } from "./components/JsonInfo";
 import { ContractTables } from "./components/tables";
 import { useContractData } from "./data";
 import { TabIndex, zContractDetailsQueryParams } from "./types";
